Only pay attention to last 36 chars of API key
We need to make sure that someone can use a service ID _and_ the new
format of API key (see next commit). If someone makes a new API key
they’ll get it in the new format, but might not want to update their
client.

The API key is always the last part of the string, so let’s ignore
anything before the last 36 characters.

This commit only does the implmentation. There don’t seem to be tests
for the client’s constructor and I’m not sure what the best way of
setting them up is. But this definitely needs tests.