Make API key include secret and service ID
In research we’ve seen people mix up the service ID and API key because
they’re both 36 character UUIDs. We can’t get rid of the service ID
because it’s used to look up the API key.

Instead, we should change API key to be one long string, which contains
both the service ID, API key and (optionally) the name of the key. For
example:

**API key**
```
casework_production-8b3aa916-ec82-434e-b0c5-d5d9b371d6a3-dcdc5083-2fee-4fba-8afd-51f3f4bcb7b0
```

This commit changes the client to make the service ID optional. If it’s
not passed, they we attempt to extract it from the secret instead.

This commit only does the implmentation. There don’t seem to be tests
for the client’s constructor and I’m not sure what the best way of
setting them up is. But this definitely needs tests.