Add a pull request template
We should clarify what our expectations are of external developers
contributing to our codebase.

A good way to do this is through a pull request template, which lets
developers check off what they’ve done. A lot of the GOV.UK frontend
repos do this, eg
https://github.com/alphagov/govuk_elements/blob/3c5c43a6fd4209cf2b6d180e7a5eadf6fa94c132/.github/PULL_REQUEST_TEMPLATE.md

This commit also documents how we do version numbering, so that an
external developer would have a reasonable chance of being able to
increment the version number in the right way.