Added precompiled letters to the Java Client.
The request only takes two keys content (which is a base64 encoded
string and the reference for the letter). Both of these are required
and must be checked at the client end. The API takes a base64 encoded
string or a Java Fil object. The Java file object must be a pdf file.

The response of a precompiled letter is different to that of a standard
letter, the subject is always static and there is no body.

* Added PDF to base64 conversion and checking
* Added new unit tests to tes the functionality
* Added Integration Tests
* Updated README.md
* Updated version number to 3.9.0